refactor(cell): move to script setup
